requirements.txt
Django
gunicorn
psycopg2-binary
dj-database-url
whitenoise
django-cors-headers
djangorestframework
djangorestframework-simplejwt
